requirements.txt
aiohappyeyeballs==2.4.0
aiohttp==3.10.5
aiosignal==1.3.1
annotated-types==0.7.0
anyascii==0.3.2
anyio==4.4.0
astor==0.8.1
attrs==24.2.0
beautifulsoup4==4.12.3
cachetools==5.5.0
certifi==2024.8.30
charset-normalizer==3.3.2
colorama==0.4.6
contourpy==1.3.0
cycler==0.12.1
Cython==3.0.11
datasets==3.0.0
decorator==5.1.1
defusedxml==0.7.1
dill==0.3.8
easyocr==1.7.1
filelock==3.16.0
fire==0.6.0
fonttools==4.53.1
frozenlist==1.4.1
fsspec==2024.6.1
google-ai-generativelanguage==0.6.9
google-api-core==2.19.2
google-api-python-client==2.145.0
google-auth==2.34.0
google-auth-httplib2==0.2.0
google-generativeai==0.8.0
googleapis-common-protos==1.65.0
grpcio==1.66.1
grpcio-status==1.66.1
h11==0.14.0
h5py==3.11.0
httpcore==1.0.5
httplib2==0.22.0
httpx==0.27.2
huggingface-hub==0.24.6
idna==3.8
imageio==2.35.1
imgaug==0.4.0
Jinja2==3.1.4
kiwisolver==1.4.7
langdetect==1.0.9
lazy_loader==0.4
lmdb==1.5.1
lxml==5.3.0
MarkupSafe==2.1.5
matplotlib==3.9.2
mpmath==1.3.0
multidict==6.1.0
multiprocess==0.70.16
networkx==3.3
ninja==1.11.1.1
numpy==1.26.4
onnx==1.16.2
opencv-contrib-python==4.10.0.84
opencv-python==4.10.0.84
opencv-python-headless==4.10.0.84
opt-einsum==3.3.0
packaging==24.1
paddleocr==2.8.1
paddlepaddle==2.6.1
pandas==2.2.2
pdf2image==1.17.0
pillow==10.4.0
proto-plus==1.24.0
protobuf==3.20.2
pyarrow==17.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.1
pyclipper==1.3.0.post5
pydantic==2.9.1
pydantic_core==2.23.3
pyparsing==3.1.4
pypdfium2==4.30.0
pytesseract==0.3.13
python-bidi==0.6.0
python-dateutil==2.9.0.post0
python-doctr==0.9.0
python-docx==1.1.2
python-dotenv==1.0.1
pytz==2024.2
PyYAML==6.0.2
rapidfuzz==3.9.7
requests==2.32.3
rsa==4.9
scikit-image==0.24.0
scipy==1.14.1
setuptools==74.1.2
shapely==2.0.6
six==1.16.0
sniffio==1.3.1
soupsieve==2.6
sympy==1.13.2
termcolor==2.4.0
tifffile==2024.8.30
torch==2.4.1
torchvision==0.19.1
tqdm==4.66.5
typing_extensions==4.12.2
tzdata==2024.1
uritemplate==4.1.1
urllib3==2.2.2
watchdog==5.0.2
xxhash==3.5.0
yarl==1.11.1
